moved ProgressDialog to new Rubberduck.UI.ParserProgress namespace
