+++ RetailCoder.VBE/Navigation/NavigateAllImplementations.cs
@@ -8,6 +8,7 @@ using Rubberduck.Parsing.Grammar;
 using Rubberduck.Parsing.Symbols;
 using Rubberduck.UI;
 using Rubberduck.UI.IdentifierReferences;
+using Rubberduck.UI.ParserProgress;
 using Rubberduck.VBEditor;
 using Rubberduck.VBEditor.VBEInterfaces.RubberduckCodePane;
