+++ RetailCoder.VBE/UI/ParserProgress/ProgressDialog.resx
